:art: rename example code to make it more clear
