+++ src/actions/getTimeEXAMPLE.ts
