+++ src/queries/useTimeEXAMPLE.ts
@@ -1,9 +1,9 @@
 import { z } from "zod";
 
-import { getTime } from "@/actions/test";
+import { getTime } from "@/actions/getTimeEXAMPLE";
 import { useSWRWithZod } from "@/utils/swr";
 
-export default function useTestQuery() {
+export default function useTime() {
   // "time" is a cache key, it can be anything you want
   return useSWRWithZod("time", getTime, z.string(), "queries/testQuery.ts");
 }
